+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improved
 * Exception messages for ConfigErrors() and PostLoad() now include the Dec name.
+* Error messages for trying to clone an object into a different object type.
 
 ### Fixed
 * Reading recorder files with objects with a missing Converter could result in an unhandled exception.